+++ app-backend/src/__tests__/handler.test.ts
@@ -8,10 +8,10 @@
 
 import {graphqlHandler, playgroundHandler} from '../handler';
 
-it('graphqlHandler should be a function', (): void => {
+it('graphqlHandler should be a function', () => {
     expect(typeof graphqlHandler).toBe('function');
 });
 
-it('playgroundHandler should be a function', (): void => {
+it('playgroundHandler should be a function', () => {
     expect(typeof playgroundHandler).toBe('function');
 });